+++ openNPL/urls.py
@@ -36,20 +36,42 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
+from django.conf.urls import url
 
 from . import views, settings
+
+from rest_framework import permissions
+from drf_yasg.views import get_schema_view
+from drf_yasg import openapi
+
+schema_view = get_schema_view(
+   openapi.Info(
+      title="openNPL API",
+      default_version='v1',
+      description="openNPL is an open source platform for the management of non-performing loans. It implements detailed European Banking Authority loan templates for NPL data",
+      terms_of_service="https://www.openriskmanagement.com/",
+      contact=openapi.Contact(email="info@openrisk.eu"),
+      license=openapi.License(name="MIT License"),
+   ),
+   public=True,
+   permission_classes=(permissions.AllowAny,),
+)
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('grappelli/', include('grappelli.urls')),  # grappelli URLS
     path('grappelli/grp_doc/', include('grappelli.urls_docs')),  # grappelli URLS
     path('', include('start.urls')),  # start URLS
-    path(r'api/', views.api_root, name='api_root'), # API root
-    path(r'api/npl_data/', include(('npl_portfolio.urls', 'npl_portfolio'), namespace='npl_portfolio')), # NPL data
+    path(r'api/', views.api_root, name='api_root'),  # API root
+    path(r'api/npl_data/', include(('npl_portfolio.urls', 'npl_portfolio'), namespace='npl_portfolio')),  # NPL data
+    url(r'^api/swagger(?P<format>\.json|\.yaml)$', schema_view.without_ui(cache_timeout=0), name='schema-json'),
+    url(r'^api/swagger/$', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
+    url(r'^api/redoc/$', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
 ]
 
 if settings.DEBUG:
     import debug_toolbar
+
     urlpatterns = [
-        path('__debug__/', include(debug_toolbar.urls)),
-    ] + urlpatterns
+                      path('__debug__/', include(debug_toolbar.urls)),
+                  ] + urlpatterns